remove some warning: array subscript has type ‘char’
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@1368 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ src/texk/web2c/luatexdir/lua/lnodelib.c
@@ -895,9 +895,9 @@ static void nodelib_pushdir (lua_State * L, int n, boolean dirnode) {
     dirstring[f++]= (n<0?'-':'+');
   }
   if (n<0) n+=64;
-  dirstring[f++] = dir_names[dir_primary[n]];
-  dirstring[f++] = dir_names[dir_secondary[n]];
-  dirstring[f++] = dir_names[dir_tertiary[n]];
+  dirstring[f++] = dir_names[(int)dir_primary[n]];
+  dirstring[f++] = dir_names[(int)dir_secondary[n]];
+  dirstring[f++] = dir_names[(int)dir_tertiary[n]];
   lua_pushstring(L,dirstring);
 }
